refactor(stripe-csa): kill stripe gateway
