+++ server/paymentProviders/stripe/index.js
@@ -7,7 +7,6 @@ import { URLSearchParams } from 'url';
 
 import models from '../../models';
 import errors from '../../lib/errors';
-import { retrieveEvent } from './gateway';
 import creditcard from './creditcard';
 import { addParamsToUrl } from '../../lib/utils';
 import stripe from '../../lib/stripe';
@@ -181,7 +180,7 @@ export default {
      * We check the event on stripe directly to be sure we don't get a fake event from
      * someone else
      */
-    return retrieveEvent({ username: requestBody.user_id }, requestBody.id).then(event => {
+    return stripe.events.retrieve(requestBody.id, { stripe_account: requestBody.user_id }).then(event => {
       if (!event || (event && !event.type)) {
         throw new errors.BadRequest('Event not found');
       }
